fix(win): MSVC build of the goroutine introspection + monotonic overflow
Two Windows regressions the self-hosted CI caught after the robustness
merge (Linux was green; Windows ran a stale .pyd that masked both):

1. pygo_introspect.c did not COMPILE on MSVC -- `__atomic_*` on a uint64_t
   field hit "error C7702: no compatible type for 'uint64_t *' in _Generic"
   (on MSVC uint64_t is `unsigned __int64`, which the plat_atomic.h shim
   doesn't name, and __atomic_fetch_add has no unsigned-long-long slot at
   all).  So the whole introspection module silently fell back to a stale
   build with none of goroutine_count/goroutines/dump/... -- and the new
   test_spawn_faultinject (which uses goroutine_count) failed.
   Fix: the goid counter + g->id are now `long long` (ids are always
   positive; long long has load/store/fetch_add slots on both compilers).

2. Windows pygo_monotonic_ns() overflowed: `QPC.ticks * 1e9 / freq` exceeds
   int64 after ~15 min uptime (10 MHz x 900 s x 1e9 > 9.2e18), wrapping to a
   negative time.  Scheduler diff-math cancelled the wrap so it went
   unnoticed, but absolute-timestamp consumers (goroutine park-age tracking)
   saw bogus values -> age always None on Windows.  Fix: overflow-safe
   sec+remainder muldiv.  Benefits all Windows timing, not just age.

Also skip the pipe-fd io-wait introspection test on Windows (pipe fds aren't
pollable by the Windows netpoll).

Validated on the win-3.13t VM: a clean rebuild now registers the
introspection API and test_introspect (22 passed, 1 skipped) +
test_gc_goroutines (4) + test_spawn_faultinject (4) all pass.  Linux
unaffected (only the Windows monotonic branch changed).

+++ src/pygo_core/pygo_introspect.h
@@ -74,7 +74,7 @@ void pygo_greg_unlink(pygo_g_t *g);
 /* Per-incarnation goroutine id, Go's goid analogue.  Contention-free:
  * a per-thread counter ORed with a per-thread base, so spawning on many
  * hubs never touches a shared cacheline.  Unique for the process life. */
-uint64_t pygo_next_goid(void);
+long long pygo_next_goid(void);
 
 /* Number of live (non-FREED) goroutines.  Takes pygo_greg_lock. */
 long pygo_goroutine_count(void);
@@ -138,7 +138,7 @@ void pygo_dump_goroutines_fd(int fd);
  * identity + Python stack come from the claim-protected
  * pygo_goroutine_frames_by_id path instead. */
 typedef struct pygo_g_info {
-    uint64_t      id;
+    long long     id;
     unsigned int  state;        /* pygo_g_state_t */
     int           park_fd;      /* netpoll fd parked on, or -1 */
     int           park_events;  /* netpoll events bitmask (1=R 2=W), or 0 */
@@ -166,7 +166,7 @@ void pygo_goroutine_snapshot_free(pygo_g_info_t *arr, long count);
  * scheduler the calling thread already owns it.  Normal interpreter
  * context only.  Returns NULL with a Python exception set only on a hard
  * error (OOM). */
-PyObject *pygo_goroutine_frames_by_id(uint64_t id);
+PyObject *pygo_goroutine_frames_by_id(long long id);
 
 #ifdef __cplusplus
 }
